Completed part 2
I forgot to write a commit for 2a and did both 2a and 2b in one go, sorry. I also wasn't to clear about what to do with 2b on the displaying part of it so I messed with the other UI properties and end up finding that mouse on click worked very well. I don't believe that was what you intended in your instructions but I will leave as is since it completes the instructions.