Lab 5 Task 1 done
I changed the saveStudentData to save the student data with a CSV and tested it along with making some tests for the saveStudentData. The actually test for the comma in between the name and grade was simply doing the application and checking the text file to see if it has a comma like its suppose to.