project 2 1A complete
I added a helper method to the mainwindoe class that makes it easier to sort. I added this helper method to where ever it was applied to automatically update the list in the desired order. I also changed the priority counter to look better since the numbers were clipping through other elements.