Completed Project 2 Part 3B
I made the Task Class an abstract class so that it can be extended to the Container Class. I also added the addTask and getSubTask to the Task class perinstruction. I made the getSubTask an abstract class and I was forced to leave the addTask as is do to compiler errors. I also tested everything to make sure it all worked.