Project 2 Part 3C completed
I made it to where when the task is clicked on the sub tasks for that task are displayed and I made a label explaining such. I was tring my best to find out how to remove the (+) from tasks that don't have a subTask but I couldn't so I'll just take the possible point deduction and won't give it anymore thought.