Completed Project 2 Part 4B
After reading the instructions properly I realize I was suppose to add the showSelectedSubTask to to mainwindow class in the last part and not just to the codebeind so I addded to this part. I made it correctly and tested before commiting my changes. I also cleaned up any and all of the checkstyle warnings minus the ones in the Task class since I need those variables to open to use ContainerTask properly.