Duplicate problem
I added a helper method to detect any duplicates in the maps so that the user will know that the contact already exists in the list and blocks the user from adding that contact again. The addContact method was tested to ensure that the helper Method works along with the addContact method as it should.